Se completa el back de borrar

+++ src/main/java/com/example/demo/PersonaService.java
@@ -6,8 +6,8 @@ import java.util.List;
 //Metodos para el proyecto crud
 public interface PersonaService {
 	public List <Persona> getAllList();
-    public Persona getById(Long id);
+    public Persona getById(long id);
 	Persona add(Persona p);
 	Persona edit(Persona p);
-	Persona delete(int id);
+	Persona delete(long id);
 }
